inc: Clean up dead code in liblights
Mostly this is just killing code that doesn't apply to the incredible.
Some functions and globals that are left get to be static, as only the
HAL_MODULE_INFO_SYM needs to be exported.  I also removed the
init_globals->mutex_init, as g_lock is declared with an initializer.

Change-Id: Icc3e5a89120e3850591838083538b9547687bbcc
Signed-off-by: Josh Stone <cuviper@gmail.com>